+++ http-verbs-play-30/src/main/scala/uk/gov/hmrc/http/HttpGet.scala
@@ -24,6 +24,7 @@ import uk.gov.hmrc.http.logging.ConnectionTracing
 
 import scala.concurrent.{ExecutionContext, Future}
 
+@deprecated("Use HttpClientV2", "15.0.0")
 trait HttpGet
   extends CoreGet
      with GetHttpTransport