+++ http-verbs-play-30/src/main/scala/uk/gov/hmrc/http/HttpPut.scala
@@ -23,6 +23,7 @@ import uk.gov.hmrc.http.logging.ConnectionTracing
 
 import scala.concurrent.{ExecutionContext, Future}
 
+@deprecated("Use HttpClientV2", "15.0.0")
 trait HttpPut
   extends CorePut
      with PutHttpTransport