+++ http-verbs-play-30/src/main/scala/uk/gov/hmrc/http/HttpReadsLegacyInstances.scala
@@ -23,16 +23,30 @@ import scala.reflect.ClassTag
 
 trait HttpReadsLegacyInstances extends HttpReadsLegacyOption with HttpReadsLegacyJson
 
-trait HttpReadsLegacyRawReads extends HttpErrorFunctions {
+private[http] object LegacyHttpErrorFunctions extends HttpErrorFunctions {
+  def handleResponse(httpMethod: String, url: String)(response: HttpResponse): HttpResponse =
+    handleResponseEither(httpMethod, url)(response)
+      .fold(err =>
+        err.statusCode match {
+          case 400 => throw new BadRequestException(badRequestMessage(httpMethod, url, response.body))
+          case 404 => throw new NotFoundException(notFoundMessage(httpMethod, url, response.body))
+          case _   => throw err
+        },
+        identity
+      )
+}
+
+trait HttpReadsLegacyRawReads {
+
   @deprecated("Use uk.gov.hmrc.http.HttpReads.Implicits instead. See README for differences.", "11.0.0")
   implicit val readRaw: HttpReads[HttpResponse] =
     (method: String, url: String, response: HttpResponse) =>
-      handleResponse(method, url)(response)
+      LegacyHttpErrorFunctions.handleResponse(method, url)(response)
 }
 
 object HttpReadsLegacyRawReads extends HttpReadsLegacyRawReads
 
-trait HttpReadsLegacyOption extends HttpErrorFunctions {
+trait HttpReadsLegacyOption {
   @deprecated("Use uk.gov.hmrc.http.HttpReads.Implicits instead. See README for differences.", "11.0.0")
   implicit def readOptionOf[P](implicit rds: HttpReads[P]): HttpReads[Option[P]] =
     (method: String, url: String, response: HttpResponse) =>
@@ -42,11 +56,11 @@ trait HttpReadsLegacyOption extends HttpErrorFunctions {
       }
 }
 
-trait HttpReadsLegacyJson extends HttpErrorFunctions {
+trait HttpReadsLegacyJson {
   @deprecated("Use uk.gov.hmrc.http.HttpReads.Implicits instead. See README for differences.", "11.0.0")
   implicit def readFromJson[O](implicit rds: Reads[O], ct: ClassTag[O]): HttpReads[O] =
     (method: String, url: String, response: HttpResponse) =>
-      readJson(method, url, handleResponse(method, url)(response).json)
+      readJson(method, url, LegacyHttpErrorFunctions.handleResponse(method, url)(response).json)
 
   @deprecated("Use uk.gov.hmrc.http.HttpReads.Implicits instead. See README for differences.", "11.0.0")
   def readSeqFromJsonProperty[O](name: String)(implicit rds: Reads[O], ct: ClassTag[O]): HttpReads[Seq[O]] =
@@ -54,7 +68,7 @@ trait HttpReadsLegacyJson extends HttpErrorFunctions {
       response.status match {
         case 204 | 404 => Seq.empty
         case _ =>
-          readJson[Seq[O]](method, url, (handleResponse(method, url)(response).json \ name).getOrElse(JsNull)) //Added JsNull here to force validate to fail - replicates existing behaviour
+          readJson[Seq[O]](method, url, (LegacyHttpErrorFunctions.handleResponse(method, url)(response).json \ name).getOrElse(JsNull)) //Added JsNull here to force validate to fail - replicates existing behaviour
       }
 
   private def readJson[A](method: String, url: String, jsValue: JsValue)(implicit rds: Reads[A], ct: ClassTag[A]): A =