+++ http-verbs-play-30/src/main/scala/uk/gov/hmrc/http/client/HttpClientV2.scala
@@ -63,14 +63,10 @@ trait RequestBuilder {
 
   /** Adds the header. If the header has already been defined (e.g. from HeaderCarrier), it will be replaced.
     * It does not affect headers not mentioned.
+    *
+    * Use `transform(_.addHttpHeaders)` to append header values to existing.
     */
   def setHeader(header: (String, String)*): RequestBuilder
-
-  @deprecated("Use setHeader", "14.5.0")
-  def replaceHeader(header: (String, String)): RequestBuilder
-
-  @deprecated("Use setHeader to add or replace, or use transform(_.addHttpHeaders) to append header values to existing", "14.5.0")
-  def addHeaders(headers: (String, String)*): RequestBuilder
 
   def withProxy: RequestBuilder